move obstacles tracking outside of lidar class

+++ include/Structure.h
@@ -37,7 +37,7 @@ struct Obstacle
 {
     static constexpr size_t kMaxPoints = 20;
     PolarPoint data[kMaxPoints] = {0, 0, 0, 0, 0};
-    int size = 0;
+    uint8_t size = 0;
 };
 
 #endif
